refactor(core): rename the kungfu.console package to kungfu.cli
The package is the CLI command-dispatch framework (commands, variants,
bridging); 'console' now reads ambiguously next to the cockpit operator
surface. Rename kungfu.console -> kungfu.cli and update every absolute import,
the bridging run_module string, and the two doc path links. find_packages
auto-discovers the renamed package, so packaging is untouched. The unrelated
runtime 'console' service location and PyInstaller's console EXE flag are left
alone.
