+++ framework/core/src/python/kungfu/cli/bridging/__registry__.py
