+++ framework/core/src/python/kungfu/cli/commands/atlas.py
@@ -9,7 +9,7 @@ import click
 import json
 import sys
 
-from kungfu.console.commands import kfc, PrioritizedCommandGroup
+from kungfu.cli.commands import kfc, PrioritizedCommandGroup
 
 atlas_command_context = kfc.pass_context()
