+++ framework/core/src/python/kungfu/cli/commands/sdk.py
@@ -13,7 +13,7 @@ import sys
 import click
 
 import kungfu
-from kungfu.console.commands import kfc
+from kungfu.cli.commands import kfc
 
 
 def _resolve_sdk_entry():
