+++ framework/core/src/python/kungfu/cli/variants/__registry__.py
